refactor: add menu toolbar
