+++ src/pages/Graphiql/DocsExplorerPanel/style.module.scss
@@ -17,7 +17,6 @@
       position: absolute;
       background: white;
       width: 100%;
-      padding: 0 var(--content-padding);
       justify-content: center;
       border-left: none;
       top: var(--header-height);
@@ -36,8 +35,8 @@
     pointer-events: all;
 
     @include sm-breakpoint {
+      width: 100%;
       left: auto;
-      border-radius: 10px;
       box-shadow: 0 5px 15px -10px var(--color-text);
     }
   }
